refactor: 서비스가 아이디만 필요한 경우에는 dto 로 변환하지마라
