+++ src/main/java/prgrms/project/stuti/domain/studygroup/controller/StudyGroupMapper.java
@@ -5,9 +5,7 @@ import lombok.NoArgsConstructor;
 import prgrms.project.stuti.domain.studygroup.controller.dto.StudyGroupCreateRequest;
 import prgrms.project.stuti.domain.studygroup.controller.dto.StudyGroupUpdateRequest;
 import prgrms.project.stuti.domain.studygroup.model.Region;
-import prgrms.project.stuti.domain.studygroup.service.dto.StudyGroupApplyDto;
 import prgrms.project.stuti.domain.studygroup.service.dto.StudyGroupCreateDto;
-import prgrms.project.stuti.domain.studygroup.service.dto.StudyGroupDeleteDto;
 import prgrms.project.stuti.domain.studygroup.service.dto.StudyGroupUpdateDto;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
@@ -40,13 +38,5 @@ public class StudyGroupMapper {
 			.imageFile(updateRequest.imageFile())
 			.description(updateRequest.description())
 			.build();
-	}
-
-	public static StudyGroupApplyDto toStudyGroupApplyDto(Long memberId, Long studyGroupId) {
-		return new StudyGroupApplyDto(memberId, studyGroupId);
-	}
-
-	public static StudyGroupDeleteDto toStudyGroupDeleteDto(Long memberId, Long studyGroupId) {
-		return new StudyGroupDeleteDto(memberId, studyGroupId);
 	}
 }
