--- src/main/java/prgrms/project/stuti/domain/studygroup/service/dto/StudyGroupApplyDto.java
@@ -1,7 +1,0 @@
-package prgrms.project.stuti.domain.studygroup.service.dto;
-
-public record StudyGroupApplyDto(
-	Long memberId,
-	Long studyGroupId
-) {
-}
