--- src/main/java/prgrms/project/stuti/domain/studygroup/service/dto/StudyGroupDeleteDto.java
@@ -1,7 +1,0 @@
-package prgrms.project.stuti.domain.studygroup.service.dto;
-
-public record StudyGroupDeleteDto(
-	Long memberId,
-	Long studyGroupId
-) {
-}
