--- src/main/java/prgrms/project/stuti/domain/studygroup/service/dto/StudyGroupIdResponse.java
@@ -1,4 +1,0 @@
-package prgrms.project.stuti.domain.studygroup.service.dto;
-
-public record StudyGroupIdResponse(Long studyGroupId) {
-}
