+++ src/main/java/prgrms/project/stuti/domain/studygroup/service/response/StudyGroupDetailResponse.java
@@ -1,4 +1,4 @@
-package prgrms.project.stuti.domain.studygroup.service.dto;
+package prgrms.project.stuti.domain.studygroup.service.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
