+++ src/main/java/prgrms/project/stuti/domain/studygroup/service/response/StudyGroupIdResponse.java
@@ -1,0 +1,4 @@
+package prgrms.project.stuti.domain.studygroup.service.response;
+
+public record StudyGroupIdResponse(Long studyGroupId) {
+}
