+++ src/main/java/prgrms/project/stuti/domain/studygroup/service/studygroup/StudyGroupService.java
@@ -17,12 +17,10 @@ import prgrms.project.stuti.domain.studygroup.model.StudyMemberRole;
 import prgrms.project.stuti.domain.studygroup.repository.PreferredMbtiRepository;
 import prgrms.project.stuti.domain.studygroup.repository.studygroup.StudyGroupRepository;
 import prgrms.project.stuti.domain.studygroup.repository.studymember.StudyMemberRepository;
-import prgrms.project.stuti.domain.studygroup.service.dto.StudyGroupApplyDto;
 import prgrms.project.stuti.domain.studygroup.service.dto.StudyGroupCreateDto;
-import prgrms.project.stuti.domain.studygroup.service.dto.StudyGroupDeleteDto;
 import prgrms.project.stuti.domain.studygroup.service.dto.StudyGroupDetailDto;
-import prgrms.project.stuti.domain.studygroup.service.dto.StudyGroupIdResponse;
-import prgrms.project.stuti.domain.studygroup.service.dto.StudyGroupDetailResponse;
+import prgrms.project.stuti.domain.studygroup.service.response.StudyGroupDetailResponse;
+import prgrms.project.stuti.domain.studygroup.service.response.StudyGroupIdResponse;
 import prgrms.project.stuti.domain.studygroup.service.dto.StudyGroupUpdateDto;
 import prgrms.project.stuti.global.error.exception.MemberException;
 import prgrms.project.stuti.global.error.exception.StudyGroupException;
@@ -71,10 +69,7 @@ public class StudyGroupService {
 	}
 
 	@Transactional
-	public StudyGroupIdResponse applyStudyGroup(StudyGroupApplyDto applyDto) {
-		Long memberId = applyDto.memberId();
-		Long studyGroupId = applyDto.studyGroupId();
-
+	public StudyGroupIdResponse applyStudyGroup(Long memberId, Long studyGroupId) {
 		validateExistingStudyMember(memberId, studyGroupId);
 		saveStudyGroupApplicant(memberId, studyGroupId);
 
@@ -82,11 +77,11 @@ public class StudyGroupService {
 	}
 
 	@Transactional
-	public StudyGroupIdResponse deleteStudyGroup(StudyGroupDeleteDto deleteDto) {
-		validateLeader(deleteDto.memberId(), deleteDto.studyGroupId());
-		updateToDeleted(deleteDto.studyGroupId());
+	public StudyGroupIdResponse deleteStudyGroup(Long memberId, Long studyGroupId) {
+		validateLeader(memberId, studyGroupId);
+		updateToDeleted(studyGroupId);
 
-		return StudyGroupConverter.toStudyGroupIdResponse(deleteDto.studyGroupId());
+		return StudyGroupConverter.toStudyGroupIdResponse(studyGroupId);
 	}
 
 	private StudyGroup saveStudyGroup(StudyGroupCreateDto createDto, String imageUrl, String thumbnailUrl) {
